Move logic for showing tbdev upload button into the component itself.
