+++ tensorboard/webapp/header/header_component.ts
@@ -12,11 +12,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 ==============================================================================*/
-import {Component, Inject, OnInit} from '@angular/core';
-
-// A list of hostname values that will trigger the tbdev-upload-button button to
-// appear.
-const LOCAL_HOSTNAMES: string[] = ['localhost', '127.0.0.1'];
+import {Component} from '@angular/core';
 
 @Component({
   selector: 'app-header',
@@ -24,7 +20,7 @@ const LOCAL_HOSTNAMES: string[] = ['localhost', '127.0.0.1'];
     <mat-toolbar color="primary">
       <span class="brand">TensorBoard</span>
       <plugin-selector class="plugins"></plugin-selector>
-      <tbdev-upload-button *ngIf="showUpload"></tbdev-upload-button>
+      <tbdev-upload-button></tbdev-upload-button>
       <app-header-reload></app-header-reload>
       <settings-button></settings-button>
       <a
@@ -49,7 +45,7 @@ const LOCAL_HOSTNAMES: string[] = ['localhost', '127.0.0.1'];
         width: 100%;
       }
 
-      tbdev-upload-button {
+      tbdev-upload-button.shown {
         margin: 0 8px 0 16px;
       }
 
@@ -71,12 +67,4 @@ const LOCAL_HOSTNAMES: string[] = ['localhost', '127.0.0.1'];
     `,
   ],
 })
-export class HeaderComponent implements OnInit {
-  showUpload: boolean = false;
-
-  constructor(@Inject('window') private window: Window) {}
-
-  ngOnInit() {
-    this.showUpload = LOCAL_HOSTNAMES.includes(this.window.location.hostname);
-  }
-}
+export class HeaderComponent {}
