+++ tensorboard/webapp/tbdev_upload/tbdev_upload_test.ts
@@ -26,6 +26,7 @@ import {TbdevUploadButtonComponent} from './tbdev_upload_button_component';
 
 describe('tbdev upload test', () => {
   const clipboardSpy = jasmine.createSpyObj('Clipboard', ['copy']);
+  let fakeWindow: any = {};
   let overlayContainer: OverlayContainer;
 
   beforeEach(async () => {
@@ -38,9 +39,45 @@ describe('tbdev upload test', () => {
         NoopAnimationsModule,
       ],
       declarations: [TbdevUploadDialogComponent, TbdevUploadButtonComponent],
-      providers: [{provide: Clipboard, useValue: clipboardSpy}],
+      providers: [
+        {provide: Clipboard, useValue: clipboardSpy},
+        {provide: 'window', useValue: fakeWindow},
+      ],
     }).compileComponents();
     overlayContainer = TestBed.inject(OverlayContainer);
+    fakeWindow.location = {
+      hostname: 'localhost',
+    };
+  });
+
+  it('does not show upload button if hostname is not localhost', async () => {
+    fakeWindow.location.hostname = 'notlocalhost.com';
+    const fixture = TestBed.createComponent(TbdevUploadButtonComponent);
+    fixture.detectChanges();
+    await fixture.whenStable();
+
+    expect(fixture.debugElement.classes.shown).toBeUndefined();
+    expect(fixture.debugElement.children.length).toEqual(0);
+  });
+
+  it('shows upload button if hostname is localhost', async () => {
+    fakeWindow.location.hostname = 'localhost';
+    const fixture = TestBed.createComponent(TbdevUploadButtonComponent);
+    fixture.detectChanges();
+    await fixture.whenStable();
+
+    expect(fixture.debugElement.classes.shown).toBeDefined();
+    expect(fixture.debugElement.children.length).toEqual(1);
+  });
+
+  it('shows upload button if hostname is 127.0.0.1', async () => {
+    fakeWindow.location.hostname = '127.0.0.1';
+    const fixture = TestBed.createComponent(TbdevUploadButtonComponent);
+    fixture.detectChanges();
+    await fixture.whenStable();
+
+    expect(fixture.debugElement.classes.shown).toBeDefined();
+    expect(fixture.debugElement.children.length).toEqual(1);
   });
 
   it('opens a dialog when clicking on the button', async () => {
